Naively move branded types to `hash-types`
